Moved github utils to a separated class
