+++ src/main/kotlin/util.kt
@@ -57,41 +57,3 @@ fun configUpdateInterval(): Long {
 fun getAuthToken(): String {
     return readConfigSafe<AuthConfig>(ConfigType.AUTH, false)!!.botToken
 }
-
-/**
- * Will send an error in the [message]?.channel if null.
- * @return the Github token, set in [AuthConfig]
- */
-suspend fun getGithubToken(message: Message?): String? {
-    val token = readConfigSafe<AuthConfig>(ConfigType.AUTH, false)?.githubToken
-    if (token == null) message?.error("Github token not set in `${ConfigType.AUTH.configPath.substring(7)}`!")
-    return token
-}
-
-/**
- * Will send an error in the [message]?.channel if null.
- * @return the default Github user, set in [UserConfig]
- */
-suspend fun getDefaultGithubUser(message: Message?): String? {
-    val repo = readConfigSafe<UserConfig>(ConfigType.USER, false)?.defaultGithubUser
-    if (repo == null) message?.error("Default user / org not set in `${ConfigType.USER.configPath.substring(7)}`!")
-    return repo
-}
-
-/**
- * [user] and [repo] is the user/repo you want to create the issue in.
- * [token] is a Github Token with repo:public_repo checked
- */
-fun createGithubIssue(issue: Issue, user: String, repo: String, token: String) {
-    val url = "https://api.github.com/repos/$user/$repo/issues"
-    val body = Gson().toJson(issue).toRequestBody("".toMediaTypeOrNull())
-
-    val request = Request.Builder()
-        .addHeader("Accept", "application/vnd.github.v3+json")
-        .addHeader("Authorization", "token $token")
-        .url(url).post(body).build()
-
-    val response = OkHttpClient().newCall(request).execute()
-
-    println(response.body?.string())
-}
